requirements.txt
celery==5.3.1
Django
drf-spectacular
djangorestframework
django-celery-beat
django-celery-results
python-decouple
redis
requests>=2.31.0
psycopg2>=2.8.6,<2.9
